5주차 배포

+++ frontend/src/components/Common/Chip/Chip.jsx
@@ -1,17 +1,20 @@
 import React, { useState } from 'react';
-import styled, { css } from 'styled-components';
-import { colorMixin } from '@/styles/mixins';
+import styled from 'styled-components';
 import DeleteIcon from '@/assets/images/delete-icon.svg';
+import { BUTTON_COLOR_TYPE } from '@/constants/css';
 
-import Box from '@/components/Common/Box';
-import Typography from '../Typography';
+import { isFunction } from '@/util';
 
-export default function Chip({ onDelete, text, color, size = 'small' }) {
+import { Box } from '@/components/Common';
+
+export default function Chip({ onDelete = null, text, color, size = 'small' }) {
     const [isShow, setIsShow] = useState(true);
 
     const handleClick = () => {
-        setIsShow(false);
-        onDelete();
+        if (isFunction(onDelete)) {
+            setIsShow(false);
+            onDelete();
+        }
     };
 
     if (!isShow) return null;
@@ -23,33 +26,18 @@ export default function Chip({ onDelete, text, color, size = 'small' }) {
             color={color}
             size={size}
         >
-            <Typography variant="span">{text}</Typography>
-            {onDelete && <Icon src={DeleteIcon} />}
+            {text}
+            {isFunction(onDelete) && <Icon src={DeleteIcon} />}
         </StyledChip>
     );
 }
 
-const ErrorColor = css`
-    ${({ theme }) => colorMixin(theme.color.white, theme.color.red)}
-`;
-
-const SuccessColor = css`
-    ${({ theme }) => colorMixin(theme.color.white, theme.color.primary)}
-`;
-
-const ChipColorType = {
-    error: ErrorColor,
-    success: SuccessColor,
-};
-
 const StyledChip = styled(Box)`
-    height: 44px;
-    padding: 6px 16px;
-    font-size: 20px;
-    ${({ color }) => ChipColorType[color]}
+    padding: 0.25rem 0.5rem;
+    box-sizing: content-box;
     border-radius: 15px;
+    ${({ color }) => BUTTON_COLOR_TYPE[color]};
     cursor: pointer;
-    margin: 0.5rem;
 `;
 
 const Icon = styled.img`
